Added npm "grunt" script
Since GruntJS 1.0 was updated to support being run without a global
install of the grunt-cli package, an npm script can be used to run
grunt directly. This is nice because it doesn't force developers to
globally install the grunt-cli npm package if they don't want to.

To use it run the following command:
    npm run grunt